Fixes issue #112:
Add missing trigger events (release, tags). Only this enables us to filter
later inside the steps (if: startsWith(...), if: github.event.name ==
..., etc.) according to our needs.

The pipeline runs on each push to develop or master, skipping the steps
for publishing on TestPyPI/PyPI. If a tag is pushed, then additionally
the step for publishing to TestPyPI gets triggered. If a release is
created out of a tag, then also the last step for publishing on PyPI
will get triggered.